requirements.txt
aioredis==1.3.1
apns2==0.7.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.14.30
botocore==1.17.30
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
colour==0.1.5
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-mathfilters==1.0.0
django-push-notifications==2.0.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.0
install==1.3.3
jmespath==0.10.0
joblib==0.16.0
msgpack==1.0.0
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.8
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-magic==0.4.18
python-rake==1.4.5
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pywebpush==1.11.0
rake-nltk==1.0.4
razorpay==1.2.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
timeago==1.0.14
tqdm==4.48.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
whitenoise==5.1.0
zipp==3.1.0
zope.interface==5.1.0
